change folder structure to MVC pattern
